discordbot/requirements.txt
peewee>=3.16.3
sanic>=23.6.0
discord.py>=2.3.2
httpx>=0.25.0